chore: make marker less hackish
