+++ webapp/src/components/ParcelPreview/ParcelCanvas/ParcelCanvas.js
@@ -3,7 +3,8 @@ import PropTypes from 'prop-types'
 import { parcelType } from 'components/types'
 import debounce from 'lodash.debounce'
 import { buildCoordinate } from 'lib/utils'
-import { getParcelAttributes, drawMarker } from 'lib/parcelUtils'
+import { getParcelAttributes } from 'lib/parcelUtils'
+import { marker } from 'lib/marker'
 
 export default class ParcelPreview extends React.PureComponent {
   static propTypes = {
@@ -118,7 +119,7 @@ export default class ParcelPreview extends React.PureComponent {
     const { nw, se } = this.state
     const ctx = this.canvas.getContext('2d')
     ctx.clearRect(0, 0, width, height)
-    let marker = null
+    let markerCenter = null
     for (let px = nw.x; px < se.x; px++) {
       for (let py = nw.y; py < se.y; py++) {
         const cx = width / 2
@@ -138,13 +139,13 @@ export default class ParcelPreview extends React.PureComponent {
             ))
         const isCenter = px === x && py === y
         if (isCenter) {
-          marker = { x: rx, y: ry }
+          markerCenter = { x: rx, y: ry }
         }
         ctx.fillStyle = backgroundColor
         ctx.fillRect(rx - size / 2, ry - size / 2, size - 1, size - 1)
       }
     }
-    drawMarker(ctx, marker.x, marker.y, 2)
+    marker.draw(ctx, markerCenter.x, markerCenter.y, 2)
   }
 
   refCanvas = canvas => {
